#17077.Commit enrol changes.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@24091 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ enrol/src/main/java/ish/oncourse/enrol/checkout/payment/PaymentEditorController.java
@@ -157,7 +157,7 @@ public class PaymentEditorController implements PaymentEditorDelegate {
                     }
                 }
                 transaction.setTax(tax);
-                transaction.setTotal(getPaymentIn().getAmount());
+                transaction.setTotal(getPaymentIn().getAmount().toBigDecimal());
                 return transaction;
             }
         }
